+++ components/buildless-serverless/internal/controller/validator/validator_test.go
@@ -1,4 +1,4 @@
-package functionvalidator
+package validator
 
 import (
 	"fmt"
